End merging, add some docs
