+++ crates/libwebrtc-sys/build.rs
@@ -26,6 +26,7 @@ fn main() -> anyhow::Result<()> {
     println!("cargo:rustc-link-lib=dylib=amstrmid");
     println!("cargo:rustc-link-lib=dylib=msdmo");
     println!("cargo:rustc-link-lib=dylib=winmm");
+    println!("cargo:rustc-link-lib=dylib=Secur32");
 
     let src_path = PathBuf::from(env::var("CARGO_MANIFEST_DIR")?);
 
@@ -43,6 +44,7 @@ fn main() -> anyhow::Result<()> {
         .include(path.join("lib/include"))
         .include(path.join("lib/include/third_party/abseil-cpp"))
         .flag("-DWEBRTC_WIN")
+        .flag("-DNOMINMAX")
         .compile("libwebrtc-sys");
 
     println!("cargo:rerun-if-changed=cpp_src/bridge.cc");
