+++ windows/include/flutter_webrtc.h
@@ -36,7 +36,7 @@ class FlutterWebRTC {
   FlutterWebRTC(FlutterWebRTCPlugin* plugin);
   virtual ~FlutterWebRTC();
 
-  rust::cxxbridge1::Box<Webrtc> webrtc = init();
+  rust::cxxbridge1::Box<Webrtc> webrtc = Init();
 
   void HandleMethodCall(
       const flutter::MethodCall<EncodableValue>& method_call,
